+++ enterprise/cmd/frontend/auth/openidconnect/config_watch.go
@@ -2,7 +2,6 @@ package openidconnect
 
 import (
 	"context"
-	"sync"
 
 	"github.com/sourcegraph/sourcegraph/cmd/frontend/auth"
 	"github.com/sourcegraph/sourcegraph/pkg/conf"
@@ -10,75 +9,34 @@ import (
 	log15 "gopkg.in/inconshreveable/log15.v2"
 )
 
-// Start trying to populate the cache of issuer metadata (given the configured OpenID Connect issuer
-// URL) immediately upon server startup and site config changes so users don't incur the wait on the
-// first auth flow request.
-func init() {
-	providersOfType := func(ps []schema.AuthProviders) []*schema.OpenIDConnectAuthProvider {
-		var pcs []*schema.OpenIDConnectAuthProvider
-		for _, p := range ps {
-			if p.Openidconnect != nil {
-				pcs = append(pcs, p.Openidconnect)
-			}
+func getProviders() []auth.Provider {
+	var cfgs []*schema.OpenIDConnectAuthProvider
+	for _, p := range conf.Get().AuthProviders {
+		if p.Openidconnect == nil {
+			continue
 		}
-		return pcs
+		cfgs = append(cfgs, p.Openidconnect)
 	}
+	providers := make([]auth.Provider, 0, len(cfgs))
+	for _, cfg := range cfgs {
+		p := &provider{config: *cfg}
+		providers = append(providers, p)
+	}
+	return providers
+}
 
-	var (
-		init = true
-
-		mu  sync.Mutex
-		cur []*schema.OpenIDConnectAuthProvider
-		reg = map[schema.OpenIDConnectAuthProvider]auth.Provider{}
-	)
-	auth.ConfWatch(func() {
-		mu.Lock()
-		defer mu.Unlock()
-
-		// Only react when the config changes.
-		new := providersOfType(conf.Get().AuthProviders)
-		diff := diffProviderConfig(cur, new)
-		if len(diff) == 0 {
-			return
-		}
-
-		if !init {
-			log15.Info("Reloading changed OpenID Connect authentication provider configuration.")
-		}
-		updates := make(map[auth.Provider]bool, len(diff))
-		for pc, op := range diff {
-			if old, ok := reg[pc]; ok {
-				delete(reg, pc)
-				updates[old] = false
-			}
-			if op {
-				new := &provider{config: pc}
-				reg[pc] = new
-				updates[new] = true
-				go func(p *provider) {
+func init() {
+	go func() {
+		conf.Watch(func() {
+			providers := getProviders()
+			for _, p := range providers {
+				go func(p auth.Provider) {
 					if err := p.Refresh(context.Background()); err != nil {
 						log15.Error("Error prefetching OpenID Connect service provider metadata.", "error", err)
 					}
-				}(new)
+				}(p)
 			}
-		}
-		auth.UpdateProviders(updates)
-		cur = new
-		init = false
-	})
-}
-
-func diffProviderConfig(old, new []*schema.OpenIDConnectAuthProvider) map[schema.OpenIDConnectAuthProvider]bool {
-	diff := map[schema.OpenIDConnectAuthProvider]bool{}
-	for _, oldPC := range old {
-		diff[*oldPC] = false
-	}
-	for _, newPC := range new {
-		if _, ok := diff[*newPC]; ok {
-			delete(diff, *newPC)
-		} else {
-			diff[*newPC] = true
-		}
-	}
-	return diff
+			auth.UpdateProviders("openidconnect", providers)
+		})
+	}()
 }